module2-workflow-orchestration/prefect/pyproject.toml
[project]
name = "prefect-playground"
version = "1.0"
description = "Workflow orchestration with Prefect"
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "prefect>=3.1.5,<4.0",
    "prefect-gcp>=0.6.2,<1.0",
    "prefect-sqlalchemy>=0.5.2,<1.0",
    "pandas>=2.2.3,<3.0",
    "pyarrow>=18.1.0,<19",
    "psycopg[binary]>=3.2.3,<4.0",
    "sqlalchemy>=2.0.36,<3.0",
    "hydra-core>=1.3.2,<2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.8.2",
]

[tool.ruff]
line-length = 100
lint.select = ["F", "E", "W", "I001"]